+++ R/bounds_hy.R
@@ -1,11 +1,18 @@
 ### Function for computing search interval
-bounds_hy <- function(yi, vi, zval, zcv, ext) {
+bounds_hy <- function(es, ext) 
+{
+  
+  ### Add objects to workspace
+  yi <- es$yi
+  vi <- es$vi
   
   ub <- max(yi + 1)
   
-  if(ext == FALSE) {
+  if(ext == FALSE) 
+  {
     lb <- max(yi - 38*sqrt(vi))
-  } else {
+  } else 
+  {
     ### If search interval for bisection function is too small extend interval 
     ### to the lowest possible lower bound. Since we do not evaluate the fraction 
     ### for the replication study the extended lower bound is manually selected