+++ R/puniform.R
@@ -18,13 +18,13 @@
 #' @param yi A vector of standardized effect sizes (see Details)
 #' @param vi A vector of sampling variances belonging to the standardized effect
 #' sizes (see Details)
-#' @param alpha A integer specifying the alpha level as used in primary studies
+#' @param alpha A numerical value specifying the alpha level as used in primary studies
 #' (default is 0.05, see Details).
 #' @param side A character indicating whether the effect sizes in the primary studies
 #' are in the right-tail of the distribution (i.e., positive)  or in the left-tail
 #' of the distribution (i.e., negative) (either \code{"right"} or \code{"left"})
 #' @param method A character indicating the method to be used \code{"P"} (default),
-#' \code{"LNP"}, \code{"LN1MINP"}, \code{"KS"}, or \code{"AD"})
+#' \code{"LNP"}, \code{"LN1MINP"}, \code{"KS"}, \code{"AD"}, or \code{"ML"}
 #' @param plot A logical indicating whether a plot showing the relation between
 #' observed and expected p-values has to be rendered (default is \code{TRUE})
 #'