+++ NEWS.md
@@ -1,3 +1,13 @@
+# Changes in puniform 0.2.2 (2020-06-19)
+
+- Updated and improved documentation
+
+- var_dif_fis(), var_boot_fis(), var_dif_rmd(), var_boot_rmd(), and var_pop() 
+were added to estimate the variability of the outcomes' effect size. These 
+functions can be used in a meta-regression model to correct for outcome reporting 
+bias with the CORB method.
+
+
 # Changes in puniform 0.2.1 (2019-08-23)
 
 - meta-plot() function is added to package